feat: 🎸 #206 增加清理动作指令

+++ src/clean/index.ts
@@ -9,8 +9,8 @@ const command = program
   .name('s clean')
   .usage('[options]')
   .option('--all', i18n('clean_up_the_environment'))
-  .option('--cache <dirname>', i18n('delete_the_file_under_the_cache'))
-  .option('--component <name>', i18n('remove_components'))
+  .option('--cache [dirname]', i18n('delete_the_file_under_the_cache'))
+  .option('--component [name]', i18n('remove_components'))
   .helpOption('-h, --help', i18n('display_help_for_command'))
   .addHelpCommand(false)
   .parse(process.argv);
@@ -27,34 +27,44 @@ if (process.argv.length > 2) {
   const devsappPath = path.join(componentsPath, 'devsapp.cn');
   const args = minimist(process.argv.slice(2), {
     boolean: ['all'],
-    string: ['cache', 'component'],
   });
-  console.log(args);
-
   if (args.all) {
     let files = fse.readdirSync(path.join(sPath));
     files = files.filter((item: string) => item !== 'access.yaml' && item !== 'set-config.yml');
-    console.log(files);
     files.forEach((file: string) => {
       rimraf.sync(path.join(sPath, file));
     });
   }
   if (args.cache) {
-    rimraf.sync(path.join(cachePath, args.cache));
-    if (args.cache === 'core') {
-      rimraf.sync(path.join(cachePath, '.s-core.lock'));
+    // cache 无参数
+    if (typeof args.cache === 'boolean') {
+      rimraf.sync(path.join(cachePath));
+    }
+    // cache 有参数
+    if (typeof args.cache === 'string') {
+      rimraf.sync(path.join(cachePath, args.cache));
+      if (args.cache === 'core') {
+        rimraf.sync(path.join(cachePath, '.s-core.lock'));
+      }
     }
   }
   if (args.component) {
-    let [source, name] = args.component.split('/');
-    // name 存在说明包含/, 比如devsapp/fc
-    if (name) {
-      rimraf.sync(path.join(devsappPath, name));
-      rimraf.sync(path.join(devsappPath, source, name));
-      rimraf.sync(path.join(githubPath, source, name));
-    } else {
-      rimraf.sync(path.join(devsappPath, source));
-      rimraf.sync(path.join(devsappPath, 'devsapp', source));
+    // component 无参数
+    if (typeof args.component === 'boolean') {
+      rimraf.sync(componentsPath);
+    }
+    // component 有参数
+    if (typeof args.component === 'string') {
+      let [source, name] = args.component.split('/');
+      // name 存在说明包含/, 比如devsapp/fc
+      if (name) {
+        rimraf.sync(path.join(devsappPath, name));
+        rimraf.sync(path.join(devsappPath, source, name));
+        rimraf.sync(path.join(githubPath, source, name));
+      } else {
+        rimraf.sync(path.join(devsappPath, source));
+        rimraf.sync(path.join(devsappPath, 'devsapp', source));
+      }
     }
   }
 }
